fix(routing): support nested slugs in docs
